Added Spring Security

+++ doclib-service/src/main/java/com/strelnikov/doclib/service/impl/DocVersionImpl.java
@@ -1,17 +1,25 @@
 package com.strelnikov.doclib.service.impl;
 
+import com.strelnikov.doclib.dto.DocFileDto;
 import com.strelnikov.doclib.dto.DocVersionDto;
 import com.strelnikov.doclib.model.documnets.Document;
 import com.strelnikov.doclib.model.documnets.DocumentFile;
 import com.strelnikov.doclib.model.documnets.DocumentVersion;
 import com.strelnikov.doclib.repository.DocVersionDao;
 import com.strelnikov.doclib.repository.DocumentDao;
+import com.strelnikov.doclib.service.DocFileActions;
 import com.strelnikov.doclib.service.DocVersionActions;
+import com.strelnikov.doclib.service.DocumentActions;
 import com.strelnikov.doclib.service.dtomapper.DtoMapper;
+import com.strelnikov.doclib.service.exceptions.UnitNotFoundException;
 import com.strelnikov.doclib.service.exceptions.VersionIsAlreadyExistException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+
+import java.io.FileNotFoundException;
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class DocVersionImpl implements DocVersionActions {
@@ -22,7 +30,9 @@ public class DocVersionImpl implements DocVersionActions {
     @Autowired
     private  DtoMapper dtoMapper;
     @Autowired
-    private  DocFileImpl fileAct;
+    private  DocFileActions fileAct;
+    @Autowired
+    private DocumentDao docDao;
 
 
     private boolean checkIsVersionExist(DocumentVersion documentVersion) {
@@ -35,19 +45,58 @@ public class DocVersionImpl implements DocVersionActions {
         return false;
     }
 
+    private boolean checkAreFilesExist(DocumentVersion docVersion){
+        for (DocumentFile docFile:docVersion.getFilesList()){
+            if (!fileAct.isFileExist(docFile)){
+                return false;
+            }
+        }
+        return true;
+    }
+
+
     @Override
-    public DocVersionDto saveDocVersion(DocVersionDto docVersionDto) throws VersionIsAlreadyExistException {
+    public DocVersionDto saveDocVersion(DocVersionDto docVersionDto) throws VersionIsAlreadyExistException, FileNotFoundException {
         DocumentVersion documentVersion = dtoMapper.mapDocVersion(docVersionDto);
         if(checkIsVersionExist(documentVersion)){
             throw new VersionIsAlreadyExistException(documentVersion);
         }else {
-            documentVersion = docVersionDao.insertDocVersion(documentVersion);
-            return dtoMapper.mapDocVersion(documentVersion);
+            if(checkAreFilesExist(documentVersion)) {
+                documentVersion = docVersionDao.insertDocVersion(documentVersion);
+                return dtoMapper.mapDocVersion(documentVersion);
+            }else{
+                throw new FileNotFoundException();
+            }
         }
     }
 
     @Override
     public void deleteDocVersion(int versionId) {
-        docVersionDao.deleteDocVersion(versionId);
+        DocumentVersion documentVersion = docVersionDao.loadDocVersion(versionId);
+        if(documentVersion!=null) {
+            Document document = documentVersion.getParentDocument();
+            List<DocumentVersion> docVerList = docVersionDao.getDocVersionList(document.getId());
+            for (int i = documentVersion.getVersion() + 1; i <= document.getActualVersion(); i++) {
+                docVerList.get(i).setVersion(i - 1);
+                docVersionDao.updateVersion(docVerList.get(i));
+            }
+            docVersionDao.deleteDocVersion(versionId);
+            document.setActualVersion(document.getActualVersion()-1);
+            docDao.updateDocument(document);
+        }
+    }
+
+    @Override
+    public List<DocVersionDto> getVersionList(int docId) throws UnitNotFoundException {
+        List<DocumentVersion> docVerList = docVersionDao.getDocVersionList(docId);
+        if(docVerList!=null) {
+            List<DocVersionDto> docVerDtoList = new ArrayList<>();
+            for (DocumentVersion docVer : docVerList) {
+                docVerDtoList.add(dtoMapper.mapDocVersion(docVer));
+            }
+            return docVerDtoList;
+        }else{
+            throw new UnitNotFoundException(docId);
+        }
     }
 }
